requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Data Processing
pandas==2.2.0
numpy==1.26.4
ta==0.11.0

# Async Processing
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Messaging
discord.py==2.3.2
python-telegram-bot==20.8

# Logging
structlog==24.1.0

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Development Tools
ruff==0.2.1

# Crypto Trading
ccxt==4.2.15
gate-api==6.103.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
schedule==1.2.1 